Correctly handle top-level AppId element
Fixes 7738
